requirements.txt
# TIDE Automation Requirements
# Install with: pip install -r requirements.txt

# API Libraries
anthropic>=0.18.0
openai>=0.27.0  # Works with both old and new versions
google-generativeai>=0.3.0

# Data Analysis
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
python-dotenv>=0.19.0
requests>=2.26.0

# Optional but recommended
tqdm>=4.62.0  # Progress bars
colorama>=0.4.4  # Colored terminal output